+++ Engine_SOURCE/yaBoneCollider.cpp
@@ -1,3 +1,4 @@
+
 #include "yaBoneCollider.h"
 #include "yaMeshData.h"
 #include "yaCollider2D.h"
@@ -38,7 +39,7 @@ namespace ya
                     bool active = false;
                     for (size_t i = 0; i < mAnimationColActiveFrame[clip].size(); i++)
                     {
-                        if (mAnimationColActiveFrame[clip][i].start< frame && mAnimationColActiveFrame[clip][i].fin > frame)
+                        if (mAnimationColActiveFrame[clip][i].start< (UINT)frame && mAnimationColActiveFrame[clip][i].fin > (UINT)frame)
                             active = true;
 
                     }
@@ -101,7 +102,7 @@ namespace ya
         SetName(L"Weapon");
         Transform* tr = GetComponent<Transform>();
         Transform* meshTr = meshObject->GetComponent<Transform>();
-        tr->SetScale(Vector3(3, 0.3, 0.3));
+        tr->SetScale(Vector3(3.f, 0.3f, 0.3f));
         tr->SetParent(meshTr);
 
         mAnimationOffsets.resize(mMeshData->GetAnimationClipCount());
